Remove duplicate impl
