+++ tinybrain/src/katago/mod.rs
@@ -1,10 +1,12 @@
 use crate::*;
 use crossbeam_channel::{select, Receiver, Sender};
 use json::*;
+use micro_model_moves::*;
 use std::convert::TryFrom;
 use std::io::{BufRead, BufReader, Write};
 use std::process::{Child, Command, Stdio};
 use std::thread;
+use uuid::Uuid;
 
 pub mod json;
